Moved unit tests for GetKanjiNotes into KanjiNoteServiceTests.
